video/requirements.txt
customtkinter==5.2.2
cv2_enumerate_cameras==1.1.15
gfpgan==1.3.8
insightface==0.7.3
numpy>=1.23.5,<2
onnx==1.16.0
onnxruntime-gpu==1.16.3; sys_platform != 'darwin'
onnxruntime-silicon==1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'
pillow==9.5.0
protobuf==4.23.2
tensorflow;
torch==2.0.1+cu118; sys_platform != 'darwin'
torch==2.0.1; sys_platform == 'darwin'
tqdm==4.66.4